requirements.txt
# Instructions:
#
# It is necessary to install python3.6-dev and memcached before:
# sudo apt-get install python3.6-dev memcached
#
# Use "pip install -r requirements.txt" to install the required packages in your environment.
#
#########################################################################################

mysqlclient==1.3.12
python-memcached==1.58
Django==1.11.7
djangorestframework==3.7.3
django-filter==1.0.4
coreapi==2.2.4
